Рефакторинг Плеера ВК. Вынесение в папку player. +AppVKTrackFunction
