+++ mobile/src/components/vk/player/AppVkPlayer.vue
@@ -17,56 +17,23 @@
                 <p class = "title">{{ track.info.title }}</p>
                 <p class = "artist">{{ track.info.artist }}</p>
             </div>
-            <div class = "trackFunction">
-                <div class = "navigation">
-                    <div class = "prev"
-                        v-on:click = "prevTrack"
-                    >
-                        <img src = "../../../static/img/ico/audio/prev.svg">
-                    </div>
-                    <div class = "activate"
-                        v-on:click = "clickActivate"
-                    >
-                        <img
-                            :src = "getIsActive.src"
-                        >
-                    </div>
-                    <div class = "next"
-                        v-on:click = "nextTrack"
-                    >
-                        <img src = "../../../static/img/ico/audio/next.svg">
-                    </div>
-                </div>
-                <div class = "activity-block">
-                    <div class = "random short-active"
-                        v-on:click = "randomTrack"
 
-                        >
-                        <img src = "../../../static/img/ico/arrow/random.svg">
-                    </div>
-                    <div class = "loop short-active"
-                        v-on:click = "loopTrack"
-                        :class = "{active: loop}"
-                    >
-                        <img src = "../../../static/img/ico/arrow/loop.svg">
-                    </div>
-                </div>
-            </div>
+            <app-vk-track-function/>
         </div>
 </template>
 
 <script>
-    import AppVkTimeLine from "../AppVkTimeLine";
+    import AppVkTimeLine from "./AppVkTimeLine";
     import Player from "../../../static/js/player";
-    import {mapState, mapMutations, mapGetters} from 'vuex';
+    import {mapState} from 'vuex';
+    import AppVkTrackFunction from "./AppVkTrackFunction";
 
     export default {
         name: "AppVkPlayer",
-        components: {AppVkTimeLine},
+        components: {AppVkTrackFunction, AppVkTimeLine},
         data:function(){
             return {
-                random:false,
-                loop:false,
+
             };
         },
         computed:{
@@ -75,37 +42,8 @@
                     return Player.getTrack(state.vk.position.index);
                 },
             }),
-            ...mapGetters({
-                getIsActive: 'vk/getIsActive',
-            }),
-        },
-        methods:{
-            ...mapMutations({
-                activatePlayer:'vk/activatePlayer',
-                updatePositionIndex: 'vk/updatePositionIndex',
-            }),
-            clickActivate: function(){
-                this.activatePlayer();
-                Player.activate();
-            },
-            prevTrack: function(){
-                Player.prev();
-                this.updatePositionIndex();
-            },
-            nextTrack: function(){
-                Player.next();
-                this.updatePositionIndex();
-            },
-            randomTrack: function(){},
-            loopTrack: function(){
-                Player.loop();
-                this.loop = Player.track.loop;
 
-                // eslint-disable-next-line no-console
-                console.log(this.loop);
-            },
         },
-
     }
 </script>
 
@@ -149,19 +87,6 @@
         border-radius: 15px;
     }
 
-    #vk-player>.trackFunction{
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-    }
-    #vk-player>.trackFunction>.activity-block{
-        padding: 10px;
-        width: 100%;
-
-        display: flex;
-        justify-content: space-between;
-    }
-
     #vk-player>.trackInfo{
         display: flex;
         flex-direction: column;
@@ -185,61 +110,4 @@
         text-overflow: ellipsis;
     }
 
-    #vk-player>.trackFunction>.navigation{
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-
-        width: 100%;
-
-        max-width: 260px;
-        padding: 5px 0 10px 0;
-    }
-    #vk-player>.trackFunction>.navigation>div{
-        height: 65px;
-        width: 65px;
-
-        border-radius: 50%;
-
-        display: flex;
-        justify-content: center;
-        align-items: center;
-
-        transition: var(--medium);
-    }
-    #vk-player>.trackFunction>.navigation>div:active{
-        background-color: var(--black3);
-    }
-    #vk-player>.trackFunction>.navigation>.prev:active>img{
-        transform: translateX(-3px);
-    }
-    #vk-player>.trackFunction>.navigation>.next:active>img{
-        transform: translateX(3px);
-    }
-    #vk-player>.trackFunction>.navigation>.activate:active>img{
-        height: 37px !important;
-    }
-    #vk-player>.trackFunction>.navigation img{
-        height: 40px;
-
-        transition: var(--medium);
-    }
-
-    #vk-player .activity-block>.short-active{
-        display: flex;
-        align-items: center;
-        justify-content: center;
-
-        height: 35px;
-        width: 35px;
-        border-radius: 50%;
-
-        transition: var(--slow);
-    }
-    #vk-player .activity-block>.short-active>img{
-        height: 29px;
-    }
-    #vk-player .activity-block>.active{
-        background-color: var(--black4);
-    }
 </style>
